Updates from code review
