+++ src/utils/getMethodFromFileName.ts
@@ -1,0 +1,3 @@
+export default function getMethodFromFileName(fileName: string): string {
+  return fileName.split('.')[0];
+}
