+++ srcOA3/components/schemas/weather/discriminator.yml.njk
@@ -1,8 +1,8 @@
 type: object
 discriminator:
     mapping:
-        typeOne: {{ schemaRefGenerator("polymorphicmodel_one.yml.njk") }}
-        typeTwo: {{ schemaRefGenerator("polymorphicmodel_two.yml.njk") }}
+        typeOne: {{ schemaComponentRefGenerator("polymorphicmodel_one.yml.njk") }}
+        typeTwo: {{ schemaComponentRefGenerator("polymorphicmodel_two.yml.njk") }}
     propertyName: selector
 oneOf:
 - $ref: polymorphicmodel_one.yml.njk
